stat object tests
